# docs/product/generated/openapi.asbuilt.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "oshi-app API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/display-settings": {
      "get": {
        "tags": [
          "display-settings"
        ],
        "summary": "Get Display Settings",
        "operationId": "get_display_settings_display_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Display Settings Display Settings Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "display-settings"
        ],
        "summary": "Put Display Settings",
        "operationId": "put_display_settings_display_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisplaySettingsBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Put Display Settings Display Settings Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/gallery-views": {
      "get": {
        "tags": [
          "gallery-views"
        ],
        "summary": "List Gallery Views",
        "operationId": "list_gallery_views_gallery_views_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response List Gallery Views Gallery Views Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "gallery-views"
        ],
        "summary": "Create Gallery View",
        "operationId": "create_gallery_view_gallery_views_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GalleryViewCreateBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Gallery View Gallery Views Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/gallery-views/{gallery_view_id}": {
      "patch": {
        "tags": [
          "gallery-views"
        ],
        "summary": "Patch Gallery View",
        "operationId": "patch_gallery_view_gallery_views__gallery_view_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "gallery_view_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Gallery View Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GalleryViewRenameBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Patch Gallery View Gallery Views  Gallery View Id  Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "gallery-views"
        ],
        "summary": "Delete Gallery View",
        "operationId": "delete_gallery_view_gallery_views__gallery_view_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "gallery_view_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Gallery View Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Gallery View Gallery Views  Gallery View Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/exports": {
      "post": {
        "tags": [
          "exports"
        ],
        "summary": "Create Export",
        "operationId": "create_export_exports_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportCreateBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Export Exports Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/exports/{export_id}": {
      "get": {
        "tags": [
          "exports"
        ],
        "summary": "Get Export",
        "operationId": "get_export_exports__export_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "export_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Export Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Export Exports  Export Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/exports/{export_id}/file": {
      "get": {
        "tags": [
          "exports"
        ],
        "summary": "Download Export File",
        "operationId": "download_export_file_exports__export_id__file_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "export_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Export Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DisplaySettingsBody": {
        "properties": {
          "text_scale": {
            "type": "integer",
            "maximum": 7.0,
            "minimum": 1.0,
            "title": "Text Scale"
          },
          "ui_density": {
            "type": "integer",
            "maximum": 7.0,
            "minimum": 1.0,
            "title": "Ui Density"
          },
          "list_sort": {
            "type": "string",
            "enum": [
              "newest",
              "name",
              "created_at"
            ],
            "title": "List Sort"
          },
          "gallery_layout": {
            "type": "string",
            "enum": [
              "grid",
              "large",
              "list"
            ],
            "title": "Gallery Layout"
          },
          "landing_page": {
            "type": "string",
            "enum": [
              "home",
              "gallery",
              "register"
            ],
            "title": "Landing Page"
          },
          "residence_region": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Residence Region"
          },
          "timezone_override": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone Override"
          },
          "date_format_mode": {
            "type": "string",
            "enum": [
              "residence",
              "ui_locale",
              "iso"
            ],
            "title": "Date Format Mode"
          },
          "currency_code_override": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 3
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency Code Override"
          },
          "currency_format_mode": {
            "type": "string",
            "enum": [
              "residence",
              "ui_locale",
              "plain"
            ],
            "title": "Currency Format Mode"
          },
          "register_start_step": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Register Start Step"
          },
          "default_storage_location_id": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Storage Location Id"
          },
          "gallery_show_name": {
            "type": "boolean",
            "title": "Gallery Show Name",
            "default": true
          },
          "gallery_show_tags": {
            "type": "boolean",
            "title": "Gallery Show Tags",
            "default": true
          },
          "gallery_show_price": {
            "type": "boolean",
            "title": "Gallery Show Price",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "text_scale",
          "ui_density",
          "list_sort",
          "gallery_layout",
          "landing_page",
          "residence_region",
          "date_format_mode",
          "currency_format_mode",
          "register_start_step"
        ],
        "title": "DisplaySettingsBody"
      },
      "ExportCreateBody": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind",
            "description": "text | media"
          }
        },
        "type": "object",
        "required": [
          "kind"
        ],
        "title": "ExportCreateBody"
      },
...
      "GalleryViewCreateBody": {
        "properties": {
          "view_name": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "View Name"
          },
          "q": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Q"
          },
          "category_tag_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Category Tag Ids"
          },
          "storage_location_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Storage Location Ids"
          },
          "color_tag_slots": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Color Tag Slots"
          },
          "list_sort": {
            "type": "string",
            "title": "List Sort",
            "default": "newest"
          }
        },
        "type": "object",
        "required": [
          "view_name"
        ],
        "title": "GalleryViewCreateBody"
      },
      "GalleryViewRenameBody": {
        "properties": {
          "view_name": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "View Name"
          }
        },
        "type": "object",
        "required": [
          "view_name"
        ],
        "title": "GalleryViewRenameBody"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}